# openapi.yaml
openapi: 3.1.0
info:
  title: Pipeline Labs API
  version: 2.0.0
  description: |
    Pipeline Labs is an SDK-first platform for orchestrating LLM and ML training workflows
    on infrastructure owned by the customer. The API supports Supabase bearer tokens for the web app
    and server-issued API keys for SDK and automation use cases.
servers:
  - url: https://pipelinelabs-ashen.vercel.app/api
    description: Production
  - url: http://localhost:3001/api
    description: Local development
...
paths:
...
  /jobs:
    get:
      tags: [Jobs]
      summary: List training jobs
      operationId: listTrainingJobs
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Training jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrainingJob'
  /jobs/start:
    post:
      tags: [Jobs]
      summary: Start training job
      operationId: startTrainingJob
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                planId:
                  type: string
                  format: uuid
                cloudProviderId:
                  type: string
                  format: uuid
              required: [planId, cloudProviderId]
      responses:
        '200':
          description: Job started
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                    format: uuid
                  status:
                    type: string
  /jobs/{jobId}:
    get:
      tags: [Jobs]
      summary: Get training job
      operationId: getTrainingJob
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingJob'
        '404':
          $ref: '#/components/responses/NotFound'
  /jobs/{jobId}/cancel:
    post:
      tags: [Jobs]
      summary: Cancel training job
      operationId: cancelTrainingJob
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /monitoring/{jobId}/logs:
    get:
      tags: [Monitoring]
      summary: Get training logs
      operationId: getTrainingLogs
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Training logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrainingLog'
...
  /monitoring/{jobId}/status:
    get:
      tags: [Monitoring]
      summary: Get training job status summary
      operationId: getTrainingJobStatus
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Status summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingJobStatus'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase access token used by the web app.
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Server-issued API key for SDK and automation workflows.
...
  schemas:
...
    TrainingJob:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        plan_id:
          type: string
          format: uuid
        dataset_id:
          type: string
          format: uuid
        cloud_provider_id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [pending, provisioning, installing, running, completed, failed, cancelled]
        instance_id:
          type: string
          nullable: true
        instance_type:
          type: string
          nullable: true
        instance_ip:
          type: string
          nullable: true
        instance_region:
          type: string
          nullable: true
        final_metrics:
          type: object
          additionalProperties:
            type: number
        model_artifact_path:
          type: string
          nullable: true
        error_message:
          type: string
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    TrainingLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        level:
          type: string
          enum: [info, warn, error, debug]
        message:
          type: string
        metadata:
          type: object
          additionalProperties: true
...
    TrainingJobStatus:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        error_message:
          type: string
          nullable: true
        final_metrics:
          type: object
          additionalProperties:
            type: number
        instance_type:
          type: string
          nullable: true
        instance_region:
          type: string
          nullable: true
        recentLogs:
          type: array
          items:
            $ref: '#/components/schemas/TrainingLog'
        latestMetrics:
          type: array
          items:
            type: object
            properties:
              metric_name:
                type: string
              metric_value:
                type: number
              epoch:
                type: integer
              step:
                type: integer
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Resource not found